+++ Frameworks/OpenMPT/OpenMPT/common/mptPathString.cpp
@@ -22,18 +22,6 @@
 #include <tchar.h>
 #endif
 
-#if MPT_OS_WINDOWS && MPT_OS_WINDOWS_WINRT
-#if defined(__MINGW32__) || defined(__MINGW64__)
-// MinGW-w64 headers do not declare this for WinRT, which is wrong.
-extern "C" {
-WINBASEAPI DWORD WINAPI GetFullPathNameW(LPCWSTR lpFileName, DWORD nBufferLength, LPWSTR lpBuffer, LPWSTR *lpFilePart);
-#ifndef GetFullPathName
-#define GetFullPathName GetFullPathNameW
-#endif
-}
-#endif
-#endif
-
 OPENMPT_NAMESPACE_BEGIN
 
 #if MPT_OS_WINDOWS
@@ -44,7 +32,11 @@ namespace mpt
 
 RawPathString PathString::AsNativePrefixed() const
 {
-	if(path.length() <= MAX_PATH || path.substr(0, 4) == PL_("\\\\?\\"))
+#if MPT_OS_WINDOWS_WINRT && (_WIN32_WINNT < 0x0a00)
+	// For WinRT on Windows 8, there is no official wy to determine an absolute path.
+	return path;
+#else
+	if(path.length() < MAX_PATH || path.substr(0, 4) == PL_("\\\\?\\"))
 	{
 		// Path is short enough or already in prefixed form
 		return path;
@@ -59,6 +51,7 @@ RawPathString PathString::AsNativePrefixed() const
 		// Regular file: C:\foo.bar -> \\?\C:\foo.bar
 		return PL_("\\\\?\\") + absPath;
 	}
+#endif
 }
 
 
@@ -442,6 +435,8 @@ bool PathIsAbsolute(const mpt::PathString &path) {
 
 #if MPT_OS_WINDOWS
 
+#if !(MPT_OS_WINDOWS_WINRT && (_WIN32_WINNT < 0x0a00))
+
 mpt::PathString GetAbsolutePath(const mpt::PathString &path)
 {
 	DWORD size = GetFullPathName(path.AsNative().c_str(), 0, nullptr, nullptr);
@@ -456,6 +451,8 @@ mpt::PathString GetAbsolutePath(const mpt::PathString &path)
 	}
 	return mpt::PathString::FromNative(fullPathName.data());
 }
+
+#endif
 
 #ifdef MODPLUG_TRACKER