+++ Frameworks/OpenMPT/OpenMPT/common/mptPathString.h
@@ -321,8 +321,12 @@ bool PathIsAbsolute(const mpt::PathString &path);
 
 #if MPT_OS_WINDOWS
 
+#if !(MPT_OS_WINDOWS_WINRT && (_WIN32_WINNT < 0x0a00))
+
 // Returns the absolute path for a potentially relative path and removes ".." or "." components. (same as GetFullPathNameW)
 mpt::PathString GetAbsolutePath(const mpt::PathString &path);
+
+#endif
 
 #ifdef MODPLUG_TRACKER