+++ Frameworks/OpenMPT/OpenMPT/common/versionNumber.h
@@ -17,7 +17,7 @@ OPENMPT_NAMESPACE_BEGIN
 // Version definitions. The only thing that needs to be changed when changing version number.
 #define VER_MAJORMAJOR  1
 #define VER_MAJOR      29
-#define VER_MINOR      10
+#define VER_MINOR      11
 #define VER_MINORMINOR 00
 
 OPENMPT_NAMESPACE_END