+++ Frameworks/OpenMPT/OpenMPT/soundlib/Load_s3m.cpp
@@ -244,9 +244,17 @@ bool CSoundFile::ReadS3M(FileReader &file, ModLoadingFlags loadFlags)
 	case S3MFileHeader::trkScreamTracker:
 		if(fileHeader.cwtv == S3MFileHeader::trkST3_20 && fileHeader.special == 0 && (fileHeader.ordNum & 0x0F) == 0 && fileHeader.ultraClicks == 0 && (fileHeader.flags & ~0x50) == 0)
 		{
-			// MPT 1.16 and older versions of OpenMPT - Simply keep default (filter) MIDI macros
-			m_dwLastSavedWithVersion = MPT_V("1.16.00.00");
-			madeWithTracker = U_("ModPlug Tracker / OpenMPT");
+			// MPT and OpenMPT before 1.17.03.02 - Simply keep default (filter) MIDI macros
+			if((fileHeader.masterVolume & 0x80) != 0)
+			{
+				m_dwLastSavedWithVersion = MPT_V("1.16.00.00");
+				madeWithTracker = U_("ModPlug Tracker / OpenMPT 1.17");
+			} else
+			{
+				// MPT 1.0 alpha5 doesn't set the stereo flag, but MPT 1.0 beta1 does.
+				m_dwLastSavedWithVersion = MPT_V("1.00.00.00");
+				madeWithTracker = U_("ModPlug Tracker 1.0 alpha");
+			}
 			keepMidiMacros = true;
 			nonCompatTracker = true;
 			m_playBehaviour.set(kST3LimitPeriod);
@@ -255,6 +263,9 @@ bool CSoundFile::ReadS3M(FileReader &file, ModLoadingFlags loadFlags)
 			madeWithTracker = U_("Velvet Studio");
 		} else
 		{
+			// ST3.20 should only ever write ultra-click values 16, 24 and 32 (corresponding to 8, 12 and 16 in the GUI), ST3.01/3.03 should only write 0.
+			// However, we won't fingerprint these values here as it's unlikely that there is any other tracker out there disguising as ST3 and using a strange ultra-click value.
+			// Also, re-saving a file with a strange ultra-click value in ST3 doesn't fix this value unless the user manually changes it, or if it's below 16.
 			madeWithTracker = U_("Scream Tracker");
 			formatTrackerStr = true;
 			isST3 = true;
@@ -276,7 +287,7 @@ bool CSoundFile::ReadS3M(FileReader &file, ModLoadingFlags loadFlags)
 		}
 		if(fileHeader.cwtv >= S3MFileHeader::trkIT2_07 && fileHeader.reserved3 != 0)
 		{
-			// Starting from  version 2.07, IT stores the total edit time of a module in the "reserved" field
+			// Starting from version 2.07, IT stores the total edit time of a module in the "reserved" field
 			uint32 editTime = DecodeITEditTimer(fileHeader.cwtv, fileHeader.reserved3);
 
 			FileHistory hist;
@@ -402,8 +413,14 @@ bool CSoundFile::ReadS3M(FileReader &file, ModLoadingFlags loadFlags)
 	// However, this version check is missing in ST3, so any mono file with a master volume of 18 will be converted to a stereo file with master volume 32.
 	else if(fileHeader.masterVolume == 2 || fileHeader.masterVolume == (2 | 0x10))
 		m_nSamplePreAmp = 0x20;
+	else if(!(fileHeader.masterVolume & 0x7F))
+		m_nSamplePreAmp = 48;
 	else
 		m_nSamplePreAmp = std::max(fileHeader.masterVolume & 0x7F, 0x10);  // Bit 7 = Stereo (we always use stereo)
+
+	const bool isStereo = (fileHeader.masterVolume & 0x80) != 0 || m_dwLastSavedWithVersion;
+	if(!isStereo)
+		m_nSamplePreAmp = Util::muldivr_unsigned(m_nSamplePreAmp, 8, 11);
 
 	// Approximately as loud as in DOSBox and a real SoundBlaster 16
 	m_nVSTiVolume = 36;
@@ -421,7 +438,8 @@ bool CSoundFile::ReadS3M(FileReader &file, ModLoadingFlags loadFlags)
 		if(fileHeader.channels[i] != 0xFF)
 		{
 			m_nChannels = i + 1;
-			ChnSettings[i].nPan = (ctype & 8) ? 0xCC : 0x33;	// 200 : 56
+			if(isStereo)
+				ChnSettings[i].nPan = (ctype & 8) ? 0xCC : 0x33;	// 200 : 56
 		}
 		if(fileHeader.channels[i] & 0x80)
 		{
@@ -695,7 +713,7 @@ bool CSoundFile::SaveS3M(std::ostream &f) const
 	fileHeader.speed = static_cast<uint8>(Clamp(m_nDefaultSpeed, 1u, 254u));
 	fileHeader.tempo = static_cast<uint8>(Clamp(m_nDefaultTempo.GetInt(), 33u, 255u));
 	fileHeader.masterVolume = static_cast<uint8>(Clamp(m_nSamplePreAmp, 16u, 127u) | 0x80);
-	fileHeader.ultraClicks = 8;
+	fileHeader.ultraClicks = 16;
 	fileHeader.usePanningTable = S3MFileHeader::idPanning;
 
 	mpt::IO::Write(f, fileHeader);