+++ Frameworks/OpenMPT/OpenMPT/soundlib/plugins/LFOPlugin.cpp
@@ -292,7 +292,7 @@ void LFOPlugin::SaveAllParameters()
 		return;
 
 	m_pMixStruct->defaultProgram = -1;
-	m_pMixStruct->pluginData.assign(chunk.cbegin(), chunk.cend());
+	m_pMixStruct->pluginData.assign(chunk.begin(), chunk.end());
 }